refactor: change email validator to main
