+++ src/main/adapters/express/express-route-adapter.ts
@@ -1,5 +1,5 @@
 import { Response, Request } from 'express'
-import { Controller, HttpRequest } from '../../presentation/protocols'
+import { Controller, HttpRequest } from '../../../presentation/protocols'
 
 export const adaptRoute = (controller: Controller): any => {
   return async (req: Request, res: Response) => {
